backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
websockets==11.0.3
requests==2.31.0
google-cloud-translate==3.11.3
edge-tts==6.1.9
openai-whisper==20230918
